Working raycasting along curves. Woohoo!

+++ src/helpers/Util.java
@@ -42,8 +42,8 @@ public class Util {
 		return false;
 	}
 
-	public static double orientation(double x0, double y0, double x1, double y1, double x2, double y2) {
-		return Math.signum((y1 - y0) * (x2 - x1) - (y2 - y1) * (x1 - x0));
+	public static int orientation(double x0, double y0, double x1, double y1, double x2, double y2) {
+		return (int) Math.signum((y1 - y0) * (x2 - x1) - (y2 - y1) * (x1 - x0));
 	}
 
 	public static int areaComplexity(Area a) {
